build: Add ui to show balances in wallet screen

+++ shared/features/wallet/src/commonMain/kotlin/com/yral/shared/features/wallet/data/models/WalletModelsDto.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class GetBalanceResponseDto(
-    val balance: Long,
+    val balance: Double,
 )
 
 @Serializable
