+++ shared/features/wallet/src/commonMain/kotlin/com/yral/shared/features/wallet/domain/models/WalletModels.kt
@@ -5,5 +5,5 @@ data class BtcInInr(
 )
 
 data class UserBtcBalance(
-    val balanceInSats: Long,
+    val balanceInSats: Double,
 )
